package.json fixes
Upped version number so that when changes are made to npm compatibility, these
changes can be uploaded to npm as a new version. Also, edited the dependencies
to be exactly equal to = rather than greater than or equal to = so as to
minimize future behavioral changes that occur due to new versions.