Fix package.json
Moved "libcoin" to be the last package so that in case libcoin isn't on npm,
you can first run "npm install" and all packages but libcoin will be installed.
Then run "npm link libcoin" to link to the local version of libcoin. So in that
case only two steps are necessary to install txtool. When the new version of
libcoin gets on npm, the "npm link libcoin" step won't be necessary.